requirements.txt
streamlit
pandas
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
plotly
matplotlib
seaborn
openpyxl
xlsxwriter
numpy